renamed permission to name and added coments

+++ commands/revoke.js
@@ -4,6 +4,7 @@ const checker = require('./helper/checkArgs.js');
 async function modifyPermission(message, args) {
     let user, permission;
 
+    // check args
     try {
         [user, permission] = await checker.checkArguments(message, args);
     } catch (error) {
@@ -11,6 +12,7 @@ async function modifyPermission(message, args) {
         return message.channel.send(`${error}`);
     }
     
+    // grab user from db
     const query = await User.findOne({
         where: {
             uid: user.id,
@@ -18,19 +20,27 @@ async function modifyPermission(message, args) {
         }
     })
 
-    const hasP = await query.hasPermission(permission);
-    if(!hasP) {
-        return message.channel.send(`${user.user.username} does not have ${permission.permission} permissions.`)
+    // if user does not exist in db, then they have no permissions
+    if(!query) {
+        return message.channel.send(`${user.user.username} does not have any permissions.`);
     }
 
+    // check if user has the permission
+    const hasP = await query.hasPermission(permission);
+    if(!hasP) {
+        return message.channel.send(`${user.user.username} does not have ${permission.name} permissions.`)
+    }
+
+    // then remove it
     await query.removePermission(permission);
     
+    // if they have no more permissions, then remove them from db
     const pCount = await query.countPermission();
     if(pCount == 0) {
         await query.destroy()
     }
 
-    return message.channel.send(`Revoked ${permission.permission} permissions for ${user.user.username}.`);
+    return message.channel.send(`Revoked ${permission.name} permissions for ${user.user.username}.`);
 }
 
 module.exports = {
